requirements.txt
websockets
python-dotenv
pydantic
colorlog
